requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers
python-dotenv
openai
tiktoken
djoser